[IK-96] refactor: 게시글 패키지 구조 변경
