+++ src/main/java/com/kdt/instakyuram/post/domain/PostImageRepository.java
@@ -1,4 +1,4 @@
-package com.kdt.instakyuram.postImage.domain;
+package com.kdt.instakyuram.post.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
