+++ src/main/java/com/kdt/instakyuram/post/domain/PostTag.java
@@ -1,27 +1,20 @@
-package com.kdt.instakyuram.postImage.domain;
+package com.kdt.instakyuram.post.domain;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
-import com.kdt.instakyuram.post.domain.Post;
-
 @Entity
-public class PostImage {
-
+public class PostTag {
 	@Id
 	private Long id;
 
+	@JoinColumn(name = "tag_id", referencedColumnName = "id")
 	@ManyToOne
+	private Tag tag;
+
 	@JoinColumn(name = "post_id", referencedColumnName = "id")
+	@ManyToOne
 	private Post post;
-
-	private String originalFileName;
-
-	private String serverFileName;
-
-	private Long size;
-
-	private String path;
 }
