+++ src/main/java/com/kdt/instakyuram/post/domain/PostTagRepository.java
@@ -1,4 +1,4 @@
-package com.kdt.instakyuram.posttag.domain;
+package com.kdt.instakyuram.post.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
