+++ src/main/java/com/kdt/instakyuram/post/service/PostImageService.java
@@ -1,4 +1,4 @@
-package com.kdt.instakyuram.postImage.service;
+package com.kdt.instakyuram.post.service;
 
 import org.springframework.stereotype.Service;
