--- src/main/java/com/kdt/instakyuram/postImage/controller/PostImageController.java
@@ -1,7 +1,0 @@
-package com.kdt.instakyuram.postImage.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class PostImageController {
-}
