--- src/main/java/com/kdt/instakyuram/posttag/controller/PostTagController.java
@@ -1,7 +1,0 @@
-package com.kdt.instakyuram.posttag.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class PostTagController {
-}
