--- src/main/java/com/kdt/instakyuram/tag/controller/TagController.java
@@ -1,7 +1,0 @@
-package com.kdt.instakyuram.tag.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class TagController {
-}
